refactor: move thread-conditional box types to `openraft-rt` crate
Consolidate all `#[cfg(feature = "single-threaded")]` type aliases in
the `openraft-rt` crate. The `openraft` crate now re-exports these types
instead of defining them locally, eliminating duplicated conditional logic.

Changes:
- Add `BoxIterator`, `BoxAsyncOnceMut`, `BoxMaybeAsyncOnceMut`, `BoxOnce`, `BoxAny` to `openraft-rt`
- Remove local `threaded` module from `openraft/src/base/mod.rs`
- Re-export all box types from `openraft_rt` in `openraft::base`
